Revert "Merge pull request #1015 from Elbehery/20250701_migrate_pages_cobra"
This reverts commit b8ae6471117d995be3f0c7805188578cab2d125f, reversing
changes made to fb8144f80b7ec2e362861f45af62c78dd76f68de.
