+++ cmd/bbolt/command_root.go
@@ -25,7 +25,6 @@ func NewRootCommand() *cobra.Command {
 		newInfoCommand(),
 		newCompactCommand(),
 		newStatsCommand(),
-		newPagesCommand(),
 	)
 
 	return rootCmd
